Major changes
Closes #35

Refactored the whole code to work with fragments and implemented a better bottom navigation bar. The thing that remains to implemented for fragments is dark mode. I'm leaving that to you @Yajiri .
