+++ SmartCarApp/app/src/main/res/layout/fragment_settings.xml
@@ -6,31 +6,40 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     tools:context=".SettingsFragment">
 
-    <androidx.appcompat.widget.SwitchCompat
-        android:id="@+id/darkMode_switch"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:paddingLeft="25dp"
-        android:text="Dark mode"
-        android:textColor="?attr/text_color"
-        android:textSize="25dp"
-        app:switchMinWidth="70dp"
-        tools:layout_editor_absoluteX="8dp"
-        tools:layout_editor_absoluteY="108dp" />
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent">
 
-    <TextView
-        android:id="@+id/textView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:layout_marginStart="10dp"
-        android:layout_marginTop="24dp"
-        android:paddingStart="20dp"
-        android:text="Settings"
-        android:textColor="?attr/text_color"
-        android:textSize="35dp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        <TextView
+            android:id="@+id/textView"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center"
+            android:layout_marginStart="16dp"
+            android:layout_marginTop="32dp"
+            android:paddingStart="20dp"
+            android:text="Settings"
+            android:textColor="?attr/text_color"
+            android:textSize="35dp"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent" />
+
+        <androidx.appcompat.widget.SwitchCompat
+            android:id="@+id/darkMode_switch"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:paddingLeft="25dp"
+            android:text="Dark mode"
+            android:textColor="?attr/text_color"
+            android:textSize="25dp"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintHorizontal_bias="0.08"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/textView"
+            app:layout_constraintVertical_bias="0.123"
+            app:switchMinWidth="70dp" />
+    </androidx.constraintlayout.widget.ConstraintLayout>
 
 
 </FrameLayout>
