--- SmartCarApp/app/src/main/res/layout/settings.xml
@@ -1,44 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="?attr/background_color"
-    tools:context=".Settings">
-
-    <androidx.appcompat.widget.SwitchCompat
-        android:id="@+id/darkMode_switch"
-        android:layout_width="250dp"
-        android:layout_height="70dp"
-        android:paddingLeft="25dp"
-        android:text="Dark mode"
-        android:textColor="?attr/text_color"
-        android:textSize="25dp"
-        app:switchMinWidth="70dp"
-        tools:layout_editor_absoluteX="8dp"
-        tools:layout_editor_absoluteY="108dp" />
-
-    <TextView
-        android:id="@+id/textView"
-        android:layout_width="353dp"
-        android:layout_height="71dp"
-        android:layout_marginStart="10dp"
-        android:layout_marginTop="24dp"
-        android:paddingStart="20dp"
-        android:text="Settings"
-        android:textColor="?attr/text_color"
-        android:textSize="35dp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <FrameLayout
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        app:layout_constraintBottom_toTopOf="@+id/bottomNavigationView"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent">
-
-    </FrameLayout>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
